vehicle/viss/0320-subscribe-filter-mixed-success.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Information Service Spec Test: Subscribe() Test</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-helper.js"></script>

<h3> Subscribe Test: Filter:mixed test</h3>
<h3> Success Condition: subscribe with filter=Interval + Range + minChange</h3>
<h3> Subscribe Signal.Drivetrain.Transmission.Speed</h3>

<h3>
<div id="log"></div>
</h3>

<script>
var t = async_test("Subscribe test: subscribe with filter=Interval + Range + minChange");

var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

var NOTIFICATION_COUNT  = 5;

var reqId = "reqid-0320";
var path = "Signal.Drivetrain.Transmission.Speed";
var phase = PH_START;
var subId = null;

var notifCount = 0;
var notif = [];

var intervalTime   = 1000; //msec
var intervalMargin = 1;    //%
var range_above    = 100;  //km/h
var range_below    = 20;   //km/h
var minChangeVal   = 5;    //km/h

var filter_interval = {};
filter_interval.interval = intervalTime;

var filter_range = {};
var objRange = {}; objRange.above = range_above; objRange.below = range_below;
filter_range.range = objRange;

var filter_minChange = {};
filter_minChange.minChange = minChangeVal;

var filter_mixed = {};
filter_mixed.interval = intervalTime;
filter_mixed.range = objRange;
filter_mixed.minChange = minChangeVal;

vehicle.onopen = function() {

  // process messages from the server
  vehicle.onmessage = function(event) {
    var ts = getTimestamp();
    var msg = JSON.parse(event.data);

    if (phase == PH_START) {
      // nothing to do
    } else if (phase == PH_SENT_SUBSCRIBE) {
      t.step_func(function() {
        // wait for subscribe response
        var res = isSubscribeSuccessResponse(reqId, msg);
        if (res) {
          addLogMessage("subscribeSuccessResponse received. subscribeId = " + msg.subscriptionId);
          subId = msg.subscriptionId
          phase = PH_SUBSCRIBED;
        } else if (isSubscribeErrorResponse(reqId, msg)) {
          helper_terminate_failure("subscribeErrorResponse received. test failed.");
          phase = PH_END;
        }
      })();
    } else if (phase == PH_SUBSCRIBED) {
      t.step_func(function() {
        // wait subscriptionNotification
        var res = isSubscriptionNotification(subId, msg);
        if (res){
          if (notifCount > NOTIFICATION_COUNT) {
            // do unsubscribe
            addLogMessage("Send unsubscribe request");
            vehicle.send('{"action":"unsubscribe"'
                          +' ,"subscriptionId":"'+subId+'"'
                          +' ,"requestId":"'+reqId+'"'
                          +'}');
            phase = PH_SENT_UNSUBSCRIBE;
          }
          notif[notifCount] = {};
          notif[notifCount].json = msg;
          notif[notifCount].timestamp = ts;
          addLogMessage("SubscriptionNotification received. value = " + msg.value + ", timestamp = " + ts);
          notifCount++;
        } else if (isSubscriptionNotificationError(subId, msg)) {
          helper_terminate_failure("SubscriptionNotificationError received. Abort test.");
          phase = PH_END;
        }
      })();
    } else if (phase == PH_SENT_UNSUBSCRIBE) {
      t.step_func(function() {
        // wait for unsubscribe response
        var res = isUnsubscribeSuccessResponse(reqId, subId, msg);
        assert_true(res, "True if unsubscribeSuccessResponse received.");
        if (res) {
          // unsubscibe success
          addLogMessage("subscribe/unsubscribe successed");
        } else if (isUnsubscribeErrorResponse(reqId, subId, msg)) {
          // unsubscribe error
          // Test has been success already. Terminate anyway.
          addLogMessage("unsubscribeErrorResponse received");
        }
        phase = PH_UNSUBSCRIBED;

        var flgSuccess = true;

        // Interval test
        addLogMessage("Interval test");
        var flgInterval = true;
        for (var i=0; i<notifCount-1;i++) {
          var res = verifyInterval(notif[i], notif[i+1]);
          if (res == false)
            flgInterval = false;
        }
        if (flgInterval == true) {
          addLogSuccess("Interval test passed");
        } else {
          addLogFailure("Interval tst failed");
        }

        // Range test
        addLogMessage("Range test");
        var flgRange = true;
        for (var i=0; i<notifCount; i++) {
          var res = verifyRange(notif[i]);
          if (res == false)
            flgRange = false;
        }
        if (flgRange == true) {
          addLogSuccess("Range test passed");
        } else {
          addLogFailure("Range test failed");
        }

        // minChange test
        addLogMessage("minChange test");
        var flgMinChange = true;
        for (var i=0; i<notifCount-1; i++) {
          var res = verifyMinChange(notif[i], notif[i+1]);
          if (res == false)
            flgMinChange = false;
        }
        if (flgMinChange == true) {
          addLogSuccess("minChange test passed");
        } else {
          addLogFailure("minChange test failed");
        }

        if (flgInterval && flgRange && flgMinChange) {
          helper_terminate_success("Totally passed");
        } else {
          helper_terminate_failure("Some tests failed");
        }

      })();
    } else if (phase == PH_UNSUBSCRIBED) {
      //nothing to do
    }
  }

  // in case server doesn't return response
  t.step_timeout(function() {
    if (phase < PH_SUBSCRIBED) {_
      helper_terminate_failure("Finish by timeout");
    }
  }, TIME_OUT_TIME
  );

  addLogMessage("Send subscribe request: path= "+path+", filter= " + JSON.stringify(filter_mixed) + "<br>");
  vehicle.send('{"action":"subscribe","path":"'+ path +'"'
                +',"requestId":"'+reqId+'"'
                +',"filters":'+ JSON.stringify(filter_mixed) +'}');
 
  phase = PH_SENT_SUBSCRIBE;
}

// verify time diff with intervalTime
// (tolerate margin of <intervalMargin>%)
function verifyInterval( notif1, notif2) {
  var diff = notif2.timestamp - notif1.timestamp;
  if ( diff > intervalTime * ((100+intervalMargin)/100)) {
    //addLogFailure("interval: "+ intervalTime +" < timediff: "+ diff +" (margin: "+intervalMargin+"%)");
    return false;
  } else 
  if ( diff < intervalTime * ((100-intervalMargin)/100)) {
    //addLogFailure("interval: "+ intervalTime +" > timediff: "+ diff +" (margin: "+intervalMargin+"%)");
    return false;
  } else {
    //addLogSuccess("interval: "+ intervalTime +" == timediff: "+ diff +" (margin: "+intervalMargin+"%)");
    return true;
  }
}
// verify Range
function verifyRange( _notif) {
  var val = _notif.json.value;
  if ( val > range_above) {
    //addLogFailure("Range_above:"+ range_above +" < value:"+ val);
    return false;
  } else 
  if ( val < range_below) {
    //addLogFailure("Range_below:"+ range_below +" > value:"+ val);
    return false;
  } else {
    //addLogSuccess("Range_above:"+ range_above +" > value:"+ val +" > Range_below:"+range_below);
    return true;
  }
}
// verify minChange
function verifyMinChange( _notif1, _notif2) {
  var val1 = _notif1.json.value;
  var val2 = _notif2.json.value;
  var change = Math.abs(val1 - val2);
  if ( change < minChangeVal) {
    //addLogFailure("Change:"+ change +" < minChange:"+ minChangeVal);
    return false;
  } else {
    //addLogSuccess("Change:"+ change +" > minChange:"+ minChangeVal);
    return true;
  }
}

</script>
